allow changing the default resource location
There are some bits in here related to reading the current resource
definitions and updating the resource definitions would be very
similair to the code added here however...

Adding a resource turned out to be a bit different, it looks like
it uses a persistent socket requireing the client to send back
the name to be used after createing the resource place holder on
the server.

I couldn't figure out a way to just send a request to create a "new"
resource other than this 2 step process.  I intend to come back to
complete this part of the api but for my imediate purposes just
updating the location of the default is adequate.

One final note, the call to change the default location will fail
if there are other resources defined. This is good as the alternative
would be those resources would probably be deleted. But basically
this shortcut to update the default only works if it is the only
resource defined.